requirements.txt
transformers==4.21.3
torch==1.12.1
pandas==1.4.3
numpy==1.23.3
deepgram-sdk==0.4.0
textwrap3==0.9.2
asyncio==3.4.3
